requirements.txt
flask
flask-sqlalchemy
psycopg2
requests 
rich
gunicorn
python-dotenv